chore: resolve circular type
