+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from "./helpers";
 export * from "./loader";
 export * from "./deploy";
 export * from "./modal";
+export * from "./api";
